backend/requirements_workflow.txt
asgiref==3.7.2
django-cors-headers==4.1.0
djangorestframework==3.14.0
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.4
djangorestframework-simplejwt==5.3.0
PyJWT==1.7.1

coverage==7.2.7
Django==3.2.23
pillow==9.5.0
typing_extensions==4.7.1
python-dotenv==0.21.1
whitenoise==6.6.0
requests==2.32.1
cloudinary==1.40.0
django-filter==24.2
dj-database-url==2.1.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
packaging==23.2